Fixed errors and configuration

+++ src/main/java/com/lukaszszumiec/recurring_payments_api/application/dto/CreateSubscriptionCommand.java
@@ -1,3 +1,5 @@
 package com.lukaszszumiec.recurring_payments_api.application.dto;
 
-public record CreateSubscriptionCommand(Long userId, double price, int billingDayOfMonth) {}
+import java.math.BigDecimal;
+
+public record CreateSubscriptionCommand(Long userId, BigDecimal price, int billingDayOfMonth) {}
